Fail fast on path+procedure and narrow analytical TTL to int
conflict set now rejects the mutually exclusive --path/--procedure combination in PreValidateSet before any network call. ReplaceAnalyticalTimeToLiveAsync now takes int? end to end, removing the overflow-prone checked cast and aligning the data-plane and ARM implementations.

+++ CosmosDBShell/Azure.Data.Cosmos.Shell.Core/ArmCosmosResourceOperations.cs
@@ -210,7 +210,7 @@ internal sealed class ArmCosmosResourceOperations(ArmCosmosContext context) : IC
         return new ContainerAnalyticalTtlView(resource.Data.Resource.AnalyticalStorageTtl);
     }
 
-    public async Task<ContainerAnalyticalTtlView> ReplaceAnalyticalTimeToLiveAsync(string databaseName, string containerName, long? analyticalTimeToLive, CancellationToken token)
+    public async Task<ContainerAnalyticalTtlView> ReplaceAnalyticalTimeToLiveAsync(string databaseName, string containerName, int? analyticalTimeToLive, CancellationToken token)
     {
         var resource = await CosmosArmResourceProvider.GetContainerAsync(context, databaseName, containerName, token);
         var data = resource.Data.Resource;
